// src/app/shared/models/moviedetails.ts
// Generated by https://quicktype.io

export interface MovieDetails {
    id:          number;
    title:       string;
    posterUrl:   string;
    backdropUrl: string;
    rating:      number;
    overview:    string;
    tagline:     string;
    budget:      number;
    revenue:     number;
    imdbUrl:     string;
    tmdbUrl:     string;
    releaseDate: string;
    runTime:     number;
    price:       number;
    genres:      Genre[];
    trailers:    Trailer[];
    casts:       Cast[];
    reviews:     Review[];
}

export interface Cast {
    id:          number;
    name:        string;
    profilePath: string;
    character:   string;
}

export interface Genre {
    id:   number;
    name: string;
}

export interface Review {
    movieId:    number;
    userId:     number;
    rating:     number;
    reviewText: string;
}

export interface Trailer {
    id:         number;
    trailerUrl: string;
    name:       string;
}